+++ rlberry/manager/tests/test_agent_manager_seeding.py
@@ -2,7 +2,6 @@ from rlberry.envs.tests.test_env_seeding import get_env_trajectory, compare_traj
 from rlberry.envs import gym_make
 from rlberry.envs.classic_control import MountainCar
 from rlberry.manager import AgentManager, MultipleManagers
-from rlberry.agents.kernel_based import RSUCBVIAgent
 from rlberry.agents.torch import A2CAgent
 import gym
 import pytest
@@ -11,9 +10,6 @@ import pytest
 @pytest.mark.parametrize(
     "env, agent_class",
     [
-        ((MountainCar, {}), RSUCBVIAgent),
-        ((gym_make, {"id": "MountainCar-v0"}), RSUCBVIAgent),
-        ((gym.make, {"id": "MountainCar-v0"}), RSUCBVIAgent),
         ((MountainCar, {}), A2CAgent),
         ((gym_make, {"id": "MountainCar-v0"}), A2CAgent),
         ((gym.make, {"id": "MountainCar-v0"}), A2CAgent),
@@ -21,10 +17,10 @@ import pytest
 )
 def test_agent_manager_and_multiple_managers_seeding(env, agent_class):
     agent_manager = AgentManager(
-        agent_class, env, fit_budget=2, init_kwargs={"horizon": 10}, n_fit=6, seed=3456
+        agent_class, env, fit_budget=2, init_kwargs={}, n_fit=6, seed=3456
     )
     agent_manager_test = AgentManager(
-        agent_class, env, fit_budget=2, init_kwargs={"horizon": 10}, n_fit=6, seed=3456
+        agent_class, env, fit_budget=2, init_kwargs={}, n_fit=6, seed=3456
     )
 
     multimanagers = MultipleManagers()